% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.camel_index}
\alias{print.camel_index}
\title{Print Method for camel_index Objects}
\usage{
\method{print}{camel_index}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"camel_index"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for camel_index Objects
}
\examples{
base_year <- camel_2015
current_year <- camel_2022

result <- camel_index(base_year, current_year)
result

}
